+++ src/Command/StatusCommand.php
@@ -6,6 +6,7 @@ use Phoenix\Command\AbstractCommand;
 use Phoenix\Migration\Manager;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class StatusCommand extends AbstractCommand
@@ -13,43 +14,47 @@ class StatusCommand extends AbstractCommand
     protected function configure()
     {
         $this->setName('status')
-            ->setDescription('List of migrations already executed and list of migrations to execute');
+            ->setDescription('List of migrations already executed and list of migrations to execute')
+            ->addOption('output-format', 'f', InputOption::VALUE_REQUIRED, 'Format of the output. Available values: table, json', 'table');
         parent::configure();
     }
 
     protected function runCommand(InputInterface $input, OutputInterface $output)
     {
-        $output->writeln('');
-        $executedMigrations = $this->manager->executedMigrations();
-
-        $output->writeln('<comment>Executed migrations</comment>');
-        if (empty($executedMigrations)) {
-            $output->writeln('<info>No executed migrations</info>');
-        } else {
-            $rows = [];
-            foreach ($executedMigrations as $migration) {
-                $rows[] = [ltrim($migration['classname'], '\\'), $migration['executed_at']];
-            }
-            $this->printTable(['Class name', 'Executed at'], $rows, $output);
+        $executedMigrations = [];
+        foreach ($this->manager->executedMigrations() as $migration) {
+            $executedMigrations[] = [
+                'migration_datetime' => $migration['migration_datetime'],
+                'classname' => ltrim($migration['classname'], '\\'),
+                'executed_at' => $migration['executed_at'],
+            ];
         }
 
-        $output->writeln('');
-
-        $migrations = $this->manager->findMigrationsToExecute(Manager::TYPE_UP);
-        $output->writeln('<comment>Migrations to execute</comment>');
-        if (empty($migrations)) {
-            $output->writeln('<info>No migrations to execute</info>');
-        } else {
-            $rows = [];
-            foreach ($migrations as $migration) {
-                $rows[] = [$migration->getClassName()];
-            }
-            $this->printTable(['Class name'], $rows, $output);
+        $migrationsToExecute = [];
+        foreach ($this->manager->findMigrationsToExecute() as $migration) {
+            $migrationsToExecute[] = [
+                'migration_datetime' => $migration->getDatetime(),
+                'classname' => $migration->getClassName()
+            ];
         }
+
+        if ($input->getOption('output-format') == 'json') {
+            $output->write(json_encode(['executed_migrations' => $executedMigrations, 'migrations_to_execute' => $migrationsToExecute]));
+            exit;
+        }
+
+        $this->printTable(['Migration datetime', 'Class name', 'Executed at'], $executedMigrations, $output, 'Executed migrations', 'No executed migrations');
+        $this->printTable(['Migration datetime', 'Class name'], $migrationsToExecute, $output, 'Migrations to execute', 'No migrations to execute');
     }
 
-    private function printTable(array $headers, array $rows, OutputInterface $output)
+    private function printTable(array $headers, array $rows, OutputInterface $output, $header, $noItemsText)
     {
+        $output->writeln('');
+        $output->writeln("<comment>$header</comment>");
+        if (empty($rows)) {
+            $output->writeln("<info>$noItemsText</info>");
+            return;
+        }
         $table = new Table($output);
         $table->setHeaders($headers);
         $table->setRows($rows);